Exchange refactoring completed
