+++ src/core/exchanges/mod.rs
@@ -2,10 +2,8 @@ pub mod application_manager;
 pub mod binance;
 pub mod cancellation_token;
 pub mod common;
-pub mod exchange;
 pub mod exchange_blocker;
-pub mod exchange_create_order;
-pub mod exchange_features;
+pub mod main;
 pub mod rest_client;
 pub mod traits;
 pub mod utils;
