+++ src/core/lifecycle/bot.rs
@@ -1,5 +1,5 @@
 use crate::core::exchanges::common::ExchangeAccountId;
-use crate::core::exchanges::exchange::Exchange;
+use crate::core::exchanges::main::exchange::Exchange;
 use crate::core::settings::CoreSettings;
 use dashmap::DashMap;
 use std::sync::Arc;
