+++ tests/binance_cancel_order.rs
@@ -2,8 +2,8 @@ use chrono::Utc;
 use mmb_lib::core::exchanges::binance::binance::*;
 use mmb_lib::core::exchanges::cancellation_token::CancellationToken;
 use mmb_lib::core::exchanges::common::*;
-use mmb_lib::core::exchanges::exchange::*;
-use mmb_lib::core::exchanges::exchange_features::*;
+use mmb_lib::core::exchanges::main::exchange::*;
+use mmb_lib::core::exchanges::main::features::*;
 use mmb_lib::core::orders::order::*;
 use mmb_lib::core::settings;
 use rust_decimal_macros::*;
