+++ tests/binance_create_order.rs
@@ -3,8 +3,8 @@ use mmb_lib::core as mmb;
 use mmb_lib::core::exchanges::binance::binance::*;
 use mmb_lib::core::exchanges::cancellation_token::CancellationToken;
 use mmb_lib::core::exchanges::common::*;
-use mmb_lib::core::exchanges::exchange::*;
-use mmb_lib::core::exchanges::exchange_features::*;
+use mmb_lib::core::exchanges::main::exchange::*;
+use mmb_lib::core::exchanges::main::features::*;
 use mmb_lib::core::orders::order::*;
 use mmb_lib::core::settings;
 use rust_decimal_macros::*;
